Per #3059, changes to address SonarQube bugs and (easily fixed) code smells flagged for PR #3114.

+++ src/basic/vx_util/num_array.cc
@@ -1337,7 +1337,7 @@ ConcatString write_css_hhmmss(const NumArray &na)
    ConcatString css;
 
    for ( int i=0; i<na.n(); ++i ) {
-      css << (i == 0 ? "" : ",") << sec_to_hhmmss(na[i]);
+      css << (i == 0 ? "" : ",") << sec_to_hhmmss(nint(na[i]));
    }
 
    return css;
@@ -1373,7 +1373,7 @@ ConcatString write_css_hhmmss(const IntArray &ia)
    ConcatString css;
 
    for ( int i=0; i<ia.n(); ++i ) {
-      css << (i == 0 ? "" : ",") << sec_to_hhmmss((double) ia[i]);
+      css << (i == 0 ? "" : ",") << sec_to_hhmmss(ia[i]);
    }
 
    return css;
