+++ src/libcode/vx_statistics/pair_data_point.cc
@@ -162,7 +162,7 @@ void PairDataPoint::assign(const PairDataPoint &pd) {
          if(add_point_pair(pd.typ_sa[i].c_str(), pd.sid_sa[i].c_str(),
                pd.lat_na[i], pd.lon_na[i],
                pd.x_na[i], pd.y_na[i],
-               pd.f_lead_na[i], pd.vld_ta[i],
+               nint(pd.f_lead_na[i]), pd.vld_ta[i],
                pd.lvl_na[i], pd.elv_na[i],
                pd.f_na[i], pd.o_na[i], pd.o_qc_sa[i].c_str(),
                cpi, pd.wgt_na[i])) {
@@ -392,8 +392,10 @@ PairDataPoint PairDataPoint::subset_pairs_cnt_thresh(
          // Handle point data
          if(is_point_vx()) {
             if(out_pd.add_point_pair(typ_sa[i].c_str(), sid_sa[i].c_str(),
-                         lat_na[i], lon_na[i], x_na[i], y_na[i],
-                         f_lead_na[i], vld_ta[i], lvl_na[i], elv_na[i],
+                         lat_na[i], lon_na[i],
+                         x_na[i], y_na[i],
+                         nint(f_lead_na[i]), vld_ta[i],
+                         lvl_na[i], elv_na[i],
                          f_na[i], o_na[i], o_qc_sa[i].c_str(),
                          cpi, wgt_na[i])) {
                out_pd.set_seeps_score(seeps_mpr[i], i);
@@ -633,8 +635,10 @@ void VxPairDataPoint::add_point_obs(float *hdr_arr, const char *hdr_typ_str,
             // Weight is from the nearest grid point
             int n = three_to_one(i_msg_typ, i_mask, i_interp);
             if(!pd[n].add_point_pair(hdr_typ_str, hdr_sid_str,
-                         hdr_lat, hdr_lon, obs_x, obs_y,
-                         bad_data_int, hdr_ut, obs_lvl, obs_hgt,
+                         hdr_lat, hdr_lon,
+                         obs_x, obs_y,
+                         bad_data_int, hdr_ut,
+                         obs_lvl, obs_hgt,
                          fcst_v, obs_v, obs_qty,
                          cpi, default_weight)) {
 
@@ -992,18 +996,16 @@ void subset_wind_pairs(const PairDataPoint &pd_u, const PairDataPoint &pd_v,
             out_pd_u.add_point_pair(pd_u.typ_sa[i].c_str(), pd_u.sid_sa[i].c_str(),
                         pd_u.lat_na[i], pd_u.lon_na[i],
                         pd_u.x_na[i], pd_u.y_na[i],
-                        pd_u.f_lead_na[i], pd_u.vld_ta[i],
+                        nint(pd_u.f_lead_na[i]), pd_u.vld_ta[i],
                         pd_u.lvl_na[i], pd_u.elv_na[i],
-                        pd_u.f_na[i], pd_u.o_na[i],
-                        pd_u.o_qc_sa[i].c_str(),
+                        pd_u.f_na[i], pd_u.o_na[i], pd_u.o_qc_sa[i].c_str(),
                         u_cpi, pd_u.wgt_na[i]);
             out_pd_v.add_point_pair(pd_v.typ_sa[i].c_str(), pd_v.sid_sa[i].c_str(),
                         pd_v.lat_na[i], pd_v.lon_na[i],
                         pd_v.x_na[i], pd_v.y_na[i],
-                        pd_v.f_lead_na[i], pd_v.vld_ta[i],
+                        nint(pd_v.f_lead_na[i]), pd_v.vld_ta[i],
                         pd_v.lvl_na[i], pd_v.elv_na[i],
-                        pd_v.f_na[i], pd_v.o_na[i],
-                        pd_v.o_qc_sa[i].c_str(),
+                        pd_v.f_na[i], pd_v.o_na[i], pd_v.o_qc_sa[i].c_str(),
                         v_cpi, pd_v.wgt_na[i]);
          }
          // Handle gridded data
@@ -1070,8 +1072,10 @@ PairDataPoint subset_climo_cdf_bin(const PairDataPoint &pd,
          // Handle point data
          if(pd.is_point_vx()) {
             out_pd.add_point_pair(pd.typ_sa[i].c_str(), pd.sid_sa[i].c_str(),
-                      pd.lat_na[i], pd.lon_na[i], pd.x_na[i], pd.y_na[i],
-                      pd.f_lead_na[i], pd.vld_ta[i], pd.lvl_na[i], pd.elv_na[i],
+                      pd.lat_na[i], pd.lon_na[i],
+                      pd.x_na[i], pd.y_na[i],
+                      nint(pd.f_lead_na[i]), pd.vld_ta[i],
+                      pd.lvl_na[i], pd.elv_na[i],
                       pd.f_na[i], pd.o_na[i], pd.o_qc_sa[i].c_str(),
                       cpi, pd.wgt_na[i]);
          }
